Refactor Rates Charts - Create a ServiceObject to Rates Charts to small controller - Use internationalization to tranlate initials in currency name to portuguese
