+++ app/controllers/rates_controller.rb
@@ -1,29 +1,13 @@
 class RatesController < ApplicationController
-  before_action :set_days
+  before_action :set_currency, only: %i(index)
 
   def index
-    @chart = LazyHighCharts::HighChart.new('graph') do |f|
-      f.title(text: 'Grafico de cotação monetária')
-      f.options[:xAxis][:categories] = @days
-      f.series(
-        type: 'spline',
-        name: 'Cotação do Dolar',
-        data: set_values
-      )
-    end
+    @chart = Charts::BuildRatesCharts.call(@currency)
   end
 
   private
 
-  def set_days
-    @days = []
-    (6).downto(0) {|day| @days << "#{Date.today-day}" }
-  end
-
-  def set_values
-    currency = params[:currency].present? ? params[:currency] : 'USDBRL'
-    @days.map do |day|
-      ExternalServices::CurrencyLayer::GetHistoricalRate.call({date: day, currency: currency})
-    end
+  def set_currency
+    @currency = params[:currency].present? ? params[:currency] : 'USDBRL'
   end
 end
