+++ config/application.rb
@@ -11,6 +11,9 @@ module ExchangeRateApp
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.2
 
+    config.time_zone = 'Brasilia'
+    config.i18n.default_locale = :'pt-BR'
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
